+++ cmd/assocs_list.go
@@ -14,7 +14,7 @@ var assocsListCmd = &cobra.Command{
 	Short: "List all catalog associations",
 	Run: func(cmd *cobra.Command, args []string) {
 		current := rc.Configurations[currentConfigName]
-		client := eclient.NewEcomClient(current.FirebaseAPIKey, current.Endpoint, timeout)
+		client := eclient.New(current.Endpoint, timeout)
 		err := client.SetToken(&current)
 		if err != nil {
 			log.Fatal(err)