+++ cmd/assocs_purge.go
@@ -13,12 +13,11 @@ var assocsPurgeCmd = &cobra.Command{
 	Short: "Purge all catalog associations",
 	Run: func(cmd *cobra.Command, args []string) {
 		current := rc.Configurations[currentConfigName]
-		client := eclient.NewEcomClient(current.FirebaseAPIKey, current.Endpoint, timeout)
+		client := eclient.New(current.Endpoint, timeout)
 		err := client.SetToken(&current)
 		if err != nil {
 			log.Fatal(err)
 		}
-
 		err = client.PurgeCatalogAssocs()
 		if err != nil {
 			log.Fatal(err)