+++ cmd/catalog_purge.go
@@ -13,7 +13,7 @@ var catalogPurgeCmd = &cobra.Command{
 	Short: "Purge the entire catalog",
 	Run: func(cmd *cobra.Command, args []string) {
 		current := rc.Configurations[currentConfigName]
-		client := eclient.NewEcomClient(current.FirebaseAPIKey, current.Endpoint, timeout)
+		client := eclient.New(current.Endpoint, timeout)
 		err := client.SetToken(&current)
 		if err != nil {
 			log.Fatal(err)