+++ cmd/products_apply.go
@@ -21,7 +21,7 @@ var productsApplyCmd = &cobra.Command{
 	Args: cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		current := rc.Configurations[currentConfigName]
-		client := eclient.NewEcomClient(current.FirebaseAPIKey, current.Endpoint, timeout)
+		client := eclient.New(current.Endpoint, timeout)
 		err := client.SetToken(&current)
 		if err != nil {
 			log.Fatal(err)