+++ cmd/profiles_remove.go
@@ -11,35 +11,44 @@ import (
 	"gopkg.in/AlecAivazis/survey.v1"
 )
 
-// projectsListCmd represents the projectsList command
-var projectsRemoveCmd = &cobra.Command{
+// profilesListCmd represents the profilesList command
+var profilesRemoveCmd = &cobra.Command{
 	Use:   "remove",
-	Short: "Remove a configuration",
-	Long:  `Removes a configuration dropping the credentials`,
+	Short: "Remove a profile",
+	Long:  `Removes a profile dropping the credentials`,
 	Run: func(cmd *cobra.Command, args []string) {
+		if len(rc.Configurations) == 0 {
+			fmt.Println("No profiles")
+			os.Exit(0)
+		}
+
 		// build a slice of "Name (Endpoint)" strings
 		pl := make([]string, 0, 8)
 		for k, v := range rc.Configurations {
 			pl = append(pl, fmt.Sprintf("%s (%s)", k, v.Endpoint))
 		}
-
-		sel := promptSelectProject(pl)
+		fmt.Println(pl)
+		sel := promptSelectProfile(pl)
 		name := sel[:strings.Index(sel, "(")-1]
-		fmt.Fprintf(os.Stdout, "Project %q selected.\n", name)
+		fmt.Fprintf(os.Stdout, "Profile %q selected.\n", name)
 
 		remove := confirm(fmt.Sprintf("Are you sure you want to remove %q", name))
 		if remove {
 			p := rc.Configurations[name]
 
+			//client := eclient.New(p.Endpoint, timeout)
 			hostname, err := configmgr.URLToHostName(p.Endpoint)
-			filename := fmt.Sprintf("%s-%s", p.FirebaseAPIKey, hostname)
+			//g, _ := client.GetConfig()
+
+			fmt.Printf("%+v\n", p)
+			filename := fmt.Sprintf("%s-%s", hostname, p.DevKey[:6])
 
 			ok, err := configmgr.DeleteProject(filename)
 			if err != nil {
-				fmt.Fprintf(os.Stderr, "Warn: remove project failed: %+v\n", errors.Cause(err))
+				fmt.Fprintf(os.Stderr, "Warn: remove profile failed: %+v\n", errors.Cause(err))
 			}
 			if !ok {
-				fmt.Fprintf(os.Stderr, "Warn: remove project failed: %+v\n", err)
+				fmt.Fprintf(os.Stderr, "Warn: remove profile failed: %+v\n", err)
 			}
 
 			// delete the configuration and write the new config to the filesystem.
@@ -49,7 +58,6 @@ var projectsRemoveCmd = &cobra.Command{
 				fmt.Fprintf(os.Stderr, "write config failed: %+v", err)
 				os.Exit(1)
 			}
-
 			fmt.Fprintf(os.Stdout, "Project %q removed.\n", name)
 			os.Exit(0)
 		}
@@ -68,5 +76,5 @@ func confirm(msg string) bool {
 }
 
 func init() {
-	projectsCmd.AddCommand(projectsRemoveCmd)
+	profilesCmd.AddCommand(profilesRemoveCmd)
 }