+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## CHANGELOG
+## v0.3.1 (Fri 17 May 2019)
++ Add Web API Key to sysinfo output.
++ configs becomes profiles.
++ Add new profiles using endpoint and devkey only.
++ No longer asks for Firebase (Google) Web API Keys but instead retrieves from from the API endpoint.
+
 ## v0.3.0 (Tue 14 May 2019)
 + `ecom catalog apply <catalog.yaml>` to replace the entire catalog.
 + `ecom catalog purge` to purge the entire catalog.